src/movie_gallery_svc/requirements.txt
flask
uvicorn 
requests
fastapi-logger
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
opentelemetry-api >= 1.10.0
opentelemetry-sdk >= 1.10.0
azure-monitor-opentelemetry 
azure-identity

dapr
dapr-ext-fastapi 
cloudevents
pydantic
retry
